% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Maps the FARS observations for specified state & years}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Integer value representing the state number}

\item{year}{The year for which the observations are required}
}
\value{
Map plot showing the observations as found for the given
  state.num and year
}
\description{
Reads the CSV data for the specified year and then filters the data
for the specified state num. The filtered data is then mapped on the
plot using the Latitude and Longitude value of each observation.
}
\examples{
fars::fars_map_state(1,2013)
fars::fars_map_state(1,2014)
fars::fars_map_state(1,2015)

}
