% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads the CSV data from FARS accidents file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The string filename located at current working directory
or the path along with filename. In case an incorrect filename or path
is provided, it will prompt an error file does not exist.
It may also prompt for error when file exists but data is non-csv.#'}
}
\value{
This function returns the data with value type as tibble in case
  the file exists or prompts an error that the file does not exists

examples
fars::fars_read(make_filename(2013))
}
\description{
The function is generic which can read data from any CSV file located
at the given path and returns the data as tibble to the caller.
}
