% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads FARS CSV data for one or more years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Integer Vector representing one or more years. In case of
non-integer value, it will prompt for coercion error. In case of any
year, whose data could not be found, error will prompt as invalid year}
}
\value{
List of tibbles for each year, having variable for month & year.
  In case of error, it would return NULL

examples
fars::fars_read_years(c(2013,2015))
fars::fars_read_years(2013:2015)
}
\description{
FARS data is available in multiple CSV files, one for each year. This
function takes vector of integers representing years and reads each of
the respective files. It adds year variable for each obs and returns a
list having tibble of months and year for obs found.
}
