% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes the observations found in FARS data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Integer Vector representing one or more years. Incorrect
values for years may result in coercion error whereas for years for
which data was not found will show error for invalid year}
}
\value{
This function returns a data.frame having observations for each
  months spreaded over the years given. In case year is not found then
  the caller is been intimidated
}
\description{
Reads the data from CSV files for each of the given years and then
summarizes the observations as counts for each month and years. The
output shows month(rows) & year(cols) wise counts
}
\examples{
fars::fars_summarize_years(2013)
fars::fars_summarize_years(c(2013,2015))
fars::fars_summarize_years(2013:2015)

}
