% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Makes the year based filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year for which the filename is required. It could be any
value that could coerced into integer, else it will prompt for error.}
}
\value{
This function returns the string filename which if exists contain
  the data for the accidents during that year

examples
fars::make_filename(2021)

No export
}
\description{
The FARS accidents data is available as year based CSV file. The function
takes the year and assumes the filename convention for FARS data and
returns the filename for the given year.
}
